refactor(model): take x-value navigation from util, not a service
`AbstractTrace` and `Candlestick` imported `NavigationService` from
`@service/navigation` and each constructed one, which `.claude/rules/model.md`
forbids in one line: "Never import from `src/service/`, `src/state/`, or
`src/ui/`. The model notifies; it does not call."

Nothing in the service layer ever constructed that class — the controller does
not build it and no service references it — so it was a model helper filed in
the wrong directory. The methods the model reached for (`computeIndexAndSegment`,
`extractXValueFromPoints`, `extractXValueFromValues`, `moveToXValueInPoints`,
`moveToXValueInValues`) are pure functions over data the trace already holds:
no state, and a `dispose()` documented as a no-op. They move to
`@util/navigation` unchanged and are called as free functions.

That also removes two allocations per trace. `Candlestick` redeclared the
inherited field and constructed a second instance in its own constructor,
discarding the base one — one dead object per trace, per live-data rebuild —
and disposed only its own copy.

Behaviour is unchanged; the moved bodies are verbatim. `NavigationService`
keeps `stepTraceInSubplot`, which is not pure and is handled separately.
`test/service/navigation.test.ts` drives `LineTrace.moveToXValue` rather than
the service, so it moves to `test/model/` where the rules say it belongs.
